+++ crates/nostr/src/event/tag/mod.rs
@@ -207,6 +207,7 @@ impl Tag {
         Self::from_standardized_without_cell(TagStandard::Coordinate {
             coordinate,
             relay_url: None,
+            uppercase: false,
         })
     }